Fix NetworkOnMainThreadException in SyslogStartConverter (Issue #24)
The start() function in the SyslogStartConverter was looking up
the localhost's name, but this is problematic if it occurs on
the main thread in an Android app with StrictMode enabled. The
name is now hard-coded to "localhost".

This fixes only part of Issue #24. The SyslogConverter must be
started with the lazy flag set to true in order to defer the server
connection until the first append.