Fix auto-config bug, where only SD config was used
The ContextInitializer used a flag to check whether AndroidManifest.xml was
read successfully. The manifest does not necessarily contain a logback config,
but the flag was incorrectly used assuming that. When auto-config saw the flag,
it stopped checking for other sources since a config was assumed found. Since
the SD card was checked before the manifest, the problem was not noticed.

This change no longer uses the flag but instead validates the data parsed from
the manifest (checks whether it's non-empty).

This fixes Issue #37.