converted server support classes to use generics
The ServerListener was converted to use a type parameter for the Client
subtype.  The Client interface was stripped down and the
remote-appender-specific methods were moved to a subinterface.  These
changes allowed the ConcurrentServerRunner and ServerSocketListener to
use a generic Client and to be reworked to move all of the
remote-appender-specific code to subclasses.  With these changes the
abstract base classes and interfaces can be moved to core to support
other server types.