added socket (re)connector
This class reimplements the much of the reconnect logic of the existing
SocketAppenderBase but does so by exposing an API that is easy to
integrate in a class that needs to handle automatic outbound connection
re-establishment, and in a way that is more easily tested.  This should
be adopted into SocketAppenderBase after some operational experience is
gained with it.