replaced references to "remote logger" with "remote receiver"
In preparation for introducing the concept of a "receiver" component
(which can be either a component that listens passively for incoming
connections from remote peers or one that actively connects to a remote
peer), this commit fixes up the taxonomy in the ServerSocketAppenderBase
implementation.  Now a ServerSocketAppender will have zero or more
connected remote receiver clients.