marked SocketAppender convenience constructors as deprecated
Convenience constructors that are used to set mutable fields that have
proper accessors add little value, and they make it more complicated to
subclass.  Best practice is to use constructor arguments only to
initialize fields that are immutable or for which an accessor method
cannot be safely provided within the design constraints of the class.

The only use of these constructors by logback modules is in the
SocketMin class.  However, because these constructors were exposed as
API (public modifier) there is a chance that other code is using them,
therefore marking them as deprecated rather than removing them.