Prevent exception stack trace from showing in tag (Issue #34)
When an exception is logged, its stack trace was appearing in the
message's logcat tags, which caused the interleaving of the exception
stack trace and the intended message in logcat. This was thought to
have been fixed in Issue #34, but it still occurred in the latest
release.

This patch appends "%nopex" to LogcatAppender's tag encoder pattern
to ensure that stack traces are not appended to the logcat tags.