Run tests in isolation; Re-enable some unit tests
This re-enables some performance tests and AsyncAppenderTest's. They're
now run in isolated JVMs to avoid the test failures from race conditions
that occur when the tests are run in parallel. The side effect is longer
test times.

This updates the SureFire configuration version, and moves the config
from the package poms to the parent.