Removes redundant and forever dangling connection
The DataSourceConnectionSource.start() method opens a new connection
only to determine whether or not the discoverConnectionProperties()
should be called. This is unnecessary since that method will perform
the same check and issue a very similar warning. A more serious
problem is that the connection is never closed.

This commit removes the use of this extra connection completely. I
think this is the last issue in need of a fix before LOGBACK-376 can
be closed.