Fixed a bug in TimeBasedFileNamingAndTriggeringPolicyBase
Fixed a bug in TimeBasedFileNamingAndTriggeringPolicyBase
causing the policy to always rollover according to the
local system time and ignore the time zone passed in the
file name pattern. LOGBACK-611 introduced the
possibility to specify time zone in the file name pattern
and the file names are now correctly created according to
the specified time zone. However, rollover is still based
on the old behaviour, e.g., if the pattern is set to
/wombat/foo.{%d, UTC}, rollover will happen at midnight
of the local system time, instead of midnight in UTC. If
for example the log is created at 00:59+01:00 02/01/2015
(local system time), the rollover will happen at
00:00+01:00 03/01/2015 instead of 01:00+01:00 02/01/2015
(i.e., 00:00Z 02/01/2015).

https://github.com/qos-ch/logback/commit/077e45c8f